requirements.txt
Django>=1.5.5
south
psycopg2==2.4.5
django-admin-bootstrapped
django-floppyforms
django-guardian
django-json-field

# Dev
django-debug-toolbar
django-rosetta

transitfeed # Read Write GTFS